requirements.txt
streamlit
google-genai
gtts
speechrecognition
pydub
